requirements.txt
streamlit>=1.36,<2.0
python-dotenv>=1.0,<2.0
langchain>=0.2.12,<0.4
langchain-community>=0.2.11,<0.4
langchain-openai>=0.1.20,<0.4
langchain-google-genai>=1.0.8,<3.0
langchain-chroma>=0.1.2,<0.3
langchain-text-splitters>=0.2.2,<0.4
chromadb>=0.5.5,<0.7
pypdf>=4.3,<6.0
duckduckgo-search>=6.2,<8.0
ddgs>=9.0,<10.0
tiktoken>=0.7,<1.0
